Refactor split file
